module_2(fix): add separate component for slider buttons
