+++ eslint.config.js
@@ -23,6 +23,6 @@ export default tseslint.config(
         "warn",
         { allowConstantExport: true },
       ],
-      },
     },
+  },
 );
