+++ src/components/home-page-blocks/news/buttons-block/buttons-block.module.scss
@@ -1,0 +1,34 @@
+@use '@styles/colors';
+
+.buttons__wrapper {
+  display: flex;
+  flex-direction: row;
+  gap: 20px;
+  justify-content: flex-end;
+}
+
+.carousel__button {
+  width: 64px;
+  height: 64px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  svg {
+    width: 24px;
+    height: 24px;
+
+    path {
+      stroke: colors.$white;
+    }
+  }
+
+  &:disabled {
+    background-color: transparent;
+    border: 1px solid colors.$blue;
+
+    svg path {
+      stroke: colors.$dark;
+    }
+  }
+}
