+++ src/components/home-page-blocks/news/buttons-block/index.ts
@@ -1,0 +1,1 @@
+export * from "./buttons-block";
